revisi course page
