+++ pelatihanku_lxp_apps/lib/presentations/commons/video_lxp.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:pelatihanku_lxp_apps/presentations/commons/radio_button_lxp.dart';
-import 'package:pelatihanku_lxp_apps/presentations/screens/pelatihanku/detail_video_pelatihanku.dart';
+import 'package:pelatihanku_lxp_apps/presentations/screens/mycourse/detail_video_pelatihanku.dart';
 import 'package:pelatihanku_lxp_apps/presentations/utils/color.dart';
 import 'package:pelatihanku_lxp_apps/presentations/utils/style.dart';
