+++ pelatihanku_lxp_apps/lib/presentations/screens/account/account.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:pelatihanku_lxp_apps/presentations/screens/akun/widget/header_account.dart';
-import 'package:pelatihanku_lxp_apps/presentations/screens/akun/widget/settings_profile.dart';
+import 'package:pelatihanku_lxp_apps/presentations/screens/account/widget/header_account.dart';
+import 'package:pelatihanku_lxp_apps/presentations/screens/account/widget/settings_profile.dart';
 
 class Account extends StatelessWidget {
   const Account({super.key});
