+++ pelatihanku_lxp_apps/lib/presentations/screens/mycourse/add_comentar.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:pelatihanku_lxp_apps/presentations/screens/pelatihanku/diskusi/reply_diskusi.dart';
+import 'package:pelatihanku_lxp_apps/presentations/screens/mycourse/discussion/reply_diskusi.dart';
 import 'package:pelatihanku_lxp_apps/presentations/utils/color.dart';
 import 'package:pelatihanku_lxp_apps/presentations/utils/style.dart';
