+++ pelatihanku_lxp_apps/lib/presentations/screens/mycourse/course/screen/course_page.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+import 'package:pelatihanku_lxp_apps/presentations/screens/mycourse/course/screen/course_view.dart';
+
+class CoursePage extends StatelessWidget {
+  const CoursePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const CourseView();
+  }
+}
