+++ pelatihanku_lxp_apps/lib/presentations/screens/mycourse/discussion/only_text_diskusi.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:pelatihanku_lxp_apps/presentations/screens/pelatihanku/add_comentar.dart';
+import 'package:pelatihanku_lxp_apps/presentations/screens/mycourse/add_comentar.dart';
 import 'package:pelatihanku_lxp_apps/presentations/utils/color.dart';
 import 'package:pelatihanku_lxp_apps/presentations/utils/style.dart';
