+++ pelatihanku_lxp_apps/lib/presentations/screens/mycourse/list_pertemuan_pelatihanku.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:pelatihanku_lxp_apps/presentations/commons/list_pertemuan_lxp.dart';
-import 'package:pelatihanku_lxp_apps/presentations/screens/pelatihanku/diskusi/blank_diskusi.dart';
-import 'package:pelatihanku_lxp_apps/presentations/screens/pelatihanku/modul_pelatihanku.dart';
-import 'package:pelatihanku_lxp_apps/presentations/screens/pelatihanku/tugas/detail_tugas.dart';
+import 'package:pelatihanku_lxp_apps/presentations/screens/mycourse/discussion/blank_diskusi.dart';
+import 'package:pelatihanku_lxp_apps/presentations/screens/mycourse/modul_pelatihanku.dart';
+import 'package:pelatihanku_lxp_apps/presentations/screens/mycourse/assignment/detail_tugas.dart';
 import 'package:pelatihanku_lxp_apps/presentations/utils/color.dart';
 import 'package:pelatihanku_lxp_apps/presentations/utils/style.dart';
